카테고리 전체 조회 (즐겨찾기 개수, 댓글 개수) 카운트 설계 변병, 내가 등록한 게시물 보기 API 변경

+++ src/main/java/com/team11/backend/dto/querydto/CategoryQueryDto.java
@@ -28,13 +28,13 @@ public class CategoryQueryDto {
     private CurrentState currentState;
     private String categoryName;
     private String createAt;
-    private Long bookmarkCnt;
-    private Long commentCnt;
+    private Integer bookmarkCnt;
+    private Integer commentCnt;
     private List<ImageQueryDto> images;
     private List<BookMarkQueryDto> bookMarks;
 
     @QueryProjection
-    public CategoryQueryDto(Long postId, String username, String nickname, String address, String title, String profileImg, String content, String myItem, String exchangeItem, CurrentState currentState, String categoryName, LocalDateTime createAt, Long bookmarkCnt, Long commentCnt) {
+    public CategoryQueryDto(Long postId, String username, String nickname, String address, String title, String profileImg, String content, String myItem, String exchangeItem, CurrentState currentState, String categoryName, LocalDateTime createAt, Integer bookmarkCnt, Integer commentCnt) {
         this.postId = postId;
         this.username = username;
         this.nickname = nickname;
